questionbank/requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
Markdown==3.9
openai==2.9.0
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.11.0
zipp==3.23.0
